* commit '0de6597679f6412b03b808da9f920fdace9ed0df':
  CHI_CheckXOP: Add hash checks for NIDAQ XOP
  CHI_CheckInstallation: Shorten MIES version so that it is readable
  clean_mies_installation*: Install the NIDAQmx XOP on Windows
  CHI_CheckInstallation: Return number of errors and add test as well
  Check Installation: Drop warnings and promote warnings to errors

Original-CommitID: 85c02c51fb4057d35ae8151d39c66cee62e8905f

+++ tools/compilation-testing/clean_mies_installation.sh
@@ -31,9 +31,11 @@ fi
 case $MSYSTEM in
   MINGW*)
       IGOR_USER_FILES="$USERPROFILE/Documents/WaveMetrics/Igor Pro 7 User Files"
+      installNIDAQmxXOP=1
       ;;
     *)
       IGOR_USER_FILES="$HOME/WaveMetrics/Igor Pro 7 User Files"
+      installNIDAQmxXOP=0
       ;;
 esac
 
@@ -53,7 +55,10 @@ if [ "$installHWXOPs" = "1" ]
 then
   cp -r  "$top_level"/XOPs-IP7-64bit/*  "$xops"
   cp -r  "$top_level"/XOP-tango-IP7-64bit/* "$xops"
-  rm -f  "$xops"/NIDAQmx64.*
+  if [ "$installNIDAQmxXOP" = "0" ]
+  then
+    rm -f  "$xops"/NIDAQmx64.*
+  fi
 else
   cp -r  "$top_level"/XOPs-IP7-64bit/HDF5*  "$xops"
   cp -r  "$top_level"/XOPs-IP7-64bit/MIESUtils*  "$xops"
